requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-bcrypt
flask-cors
flask-limiter
pyjwt
marshmallow
python-dotenv
